feat:动态用户头像

+++ base/lib/entity/circle/circle_dynamic_bean.dart
@@ -10,7 +10,7 @@ class CircleDynamicBean {
   String content;
   int commentCount;
   CircleUserBean user;
-  double createdAt;
+  int createdAt;
   String objectId;
   List<CircleImageBean> images;
   CircleLinkBean webpageLink;
